[refactor] Transfer bulk action files to a directory
