+++ wagtail/admin/views/bulk_action/__init__.py
@@ -1,0 +1,5 @@
+from .base_bulk_action import BulkAction
+from .dispatcher import index
+
+
+__all__ = ['BulkAction', 'index']
